+++ src/CachePlugin.php
@@ -1,9 +1,8 @@
 <?php
 
-namespace Http\Client\Common\Plugin;
+namespace Http\Client\Common;
 
-use Http\Client\Common\Plugin;
-use Http\Client\Common\RewindStreamException;
+use Http\Client\Common\Exception\RewindStreamException;
 use Http\Message\StreamFactory;
 use Http\Promise\FulfilledPromise;
 use Psr\Cache\CacheItemInterface;
